requirements.txt
Flask==1.1.1
requests==2.7.0
PyYAML==5.3.1
pandas==1.0.3